split config file
