+++ inference_perf/config/client/modelserver/config.py
@@ -1,0 +1,34 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from enum import Enum
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class ModelServerType(Enum):
+    VLLM = "vllm"
+    SGLANG = "sglang"
+    TGI = "tgi"
+    MOCK = "mock"
+
+
+class ModelServerClientConfig(BaseModel):
+    type: ModelServerType = ModelServerType.VLLM
+    model_name: Optional[str] = None
+    base_url: str
+    ignore_eos: bool = True
+    api_key: Optional[str] = None
+    cert_path: Optional[str] = None
+    key_path: Optional[str] = None
