+++ inference_perf/config/datagen/__init__.py
@@ -1,0 +1,62 @@
+# Copyright 2026 The Kubernetes Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from inference_perf.config.datagen.config import (
+    DataConfig,
+    DataGenType,
+    SharedPrefix,
+)
+from inference_perf.config.datagen.multimodal import (
+    AnyResolution,
+    AudioDatagenConfig,
+    ImageDatagenConfig,
+    MediaDatagenConfig,
+    Resolution,
+    ResolutionPreset,
+    SyntheticMultimodalDatagenConfig,
+    VideoDatagenConfig,
+    VideoProfile,
+    WeightedDuration,
+    WeightedResolution,
+    WeightedVideoProfile,
+)
+from inference_perf.config.datagen.replay import (
+    ConversationReplayConfig,
+    OTelTraceReplayConfig,
+    SessionReplayConfig,
+    TraceConfig,
+    TraceFormat,
+)
+
+__all__ = [
+    "AnyResolution",
+    "AudioDatagenConfig",
+    "ConversationReplayConfig",
+    "DataConfig",
+    "DataGenType",
+    "ImageDatagenConfig",
+    "MediaDatagenConfig",
+    "OTelTraceReplayConfig",
+    "Resolution",
+    "ResolutionPreset",
+    "SessionReplayConfig",
+    "SharedPrefix",
+    "SyntheticMultimodalDatagenConfig",
+    "TraceConfig",
+    "TraceFormat",
+    "VideoDatagenConfig",
+    "VideoProfile",
+    "WeightedDuration",
+    "WeightedResolution",
+    "WeightedVideoProfile",
+]
